ChessCam/Assets/Backend/requirements.txt
ultralytics
fastapi
uvicorn
pillow
python-multipart
sqlalchemy
asyncpg
werkzeug